feat(serviceAccounts): enable customization of service account resource provider (#668)
introduces BaseServiceAccountResourceProvider to capture service account specific
access checks"
